refactor: address PR #106 feedback - code organization improvements
- Add artifact field to sendMessage options for consistency
- Extract DebugOverlay component to separate file
- Create useDebugSelection hook for debug logic reusability
- Maintain existing functionality while improving code organization
